+++ src/HardenedFetch.ts
@@ -14,7 +14,7 @@ const defaults: Options = {
   doNotRetryMethods: ['CONNECT', 'DELETE', 'PATCH', 'POST', 'PUT'],
   doNotRetryCodes: [400, 401, 403, 404, 405, 406, 410, 411, 412, 422, 451, 501],
   // Rate limit options
-  rateLimitHeader: 'Retry-After',
+  rateLimitHeaders: ['Retry-After', 'RateLimit-Reset', 'X-RateLimit-Reset', 'X-Rate-Limit-Reset'],
   resetFormat: 'seconds',
 }